requirements.txt
# Core
pytz==2021.3
Django==3.2.9
django-configurations==2.3.1
gunicorn==20.1.0
newrelic==7.2.4.171
python-dotenv==0.19.2

# For the persistence stores
psycopg2-binary==2.9.2
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1

# RESTful APIs
Markdown==3.3.6
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4

# Testing
mock==4.0.3
factory-boy==3.2.1
django-nose==1.4.7
nose-progressive==1.5.2
coverage==6.2

# Static and Media Storage
django-storages==1.12.3
boto3==1.20.20

